requirements.txt
asgiref==3.8.1
boto3==1.34.1
botocore==1.34.81
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0
django-crispy-forms==2.0
django-ses==3.5.2
django-storages==1.14.2
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
packaging==24.0
pdf2image==1.17.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
